runarion-stable-diffusion/requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
controlnet-aux==0.0.9
diffusers==0.33.1
einops==0.8.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
gradio==5.25.2
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.6
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semantic-version==2.10.0
setuptools==79.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tifffile==2025.3.30
timm==0.6.7
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
xformers==0.0.29.post3
zipp==3.21.0
